20230817修正分

+++ src/main/java/com/s_giken/training/webapp/controller/MemberController.java
@@ -76,7 +76,7 @@ public class MemberController {
 			Model model) {
 		var member = memberService.findById(id);
 		if (!member.isPresent()) {
-			throw new NotFoundException("");
+			throw new NotFoundException(String.format("指定したmemberId(%d)の加入者情報が存在しません。", id));
 		}
 		model.addAttribute("member", member);
 		return "member_edit";
@@ -127,6 +127,11 @@ public class MemberController {
 	public String deleteMember(
 			@PathVariable int id,
 			RedirectAttributes redirectAttributes) {
+		var member = memberService.findById(id);
+		if (!member.isPresent()) {
+			throw new NotFoundException(String.format("指定したmemberId(%d)の加入者情報が存在しません。", id));
+		}
+
 		memberService.deleteById(id);
 		redirectAttributes.addFlashAttribute("message", "削除しました。");
 		return "redirect:/member/search";
